Initial refactor of command listener, services created
