+++ src/commandListener/Init.java
@@ -1,7 +1,11 @@
 package commandListener;
 
+import commandServices.BattleService;
+import commandServices.CellService;
+import commandServices.GameService;
+import commandServices.InventoryService;
+import commandServices.MovementService;
 import uiView.classes.PlayWindow;
-import utilities.GameCommands;
 import utilities.InputParser;
 
 public class Init {
@@ -11,8 +15,9 @@ public class Init {
 	MovementListener movementListener = new MovementListener();
 	InventoryListener inventoryListener = new InventoryListener();
 	SystemListener systemListener = new SystemListener();
+	CellListener cellListener = new CellListener();
 	
-	public void initializeListeners(PlayWindow play, GameCommands system) {
+	public void initializeListeners(PlayWindow play, GameService system) {
 		
         String[] stringArray = InputParser.parse(play.requestInput());
         
@@ -25,9 +30,22 @@ public class Init {
         //check the listeners
         
         	Reply systemReply = systemListener.listen(command, system);
-        	Reply battleReply;
-        	Reply movementReply;
+        	Reply battleReply = battleListener.listen(command, parameter);
+        	Reply movementReply = movementListener.listen(command, parameter);
+        	Reply inventoryReply = inventoryListener.listen(command, parameter);
+        	Reply cellReply = cellListener.listen(command, parameter);
         	
+        	Reply[] replies = {systemReply, battleReply, movementReply, inventoryReply, cellReply};
+        	
+        	for(Reply reply: replies) {
+        		if(reply.isSuccess) {
+        			output = reply.output;
+        			if(reply.upperOutput != null) {
+        				upperOutput = reply.upperOutput;
+        			}
+        			break;
+        		}
+        	}
         	
         //sends output generated from user selection to the GUI window
         play.outGUI(output);
