--- src/commandListener/PlayerListener.java
@@ -1,5 +1,0 @@
-package commandListener;
-
-public class PlayerListener {
-
-}
