--- src/gameplay/CellService.java
@@ -1,5 +1,0 @@
-package gameplay;
-
-public class CellService {
-
-}
