+++ src/utilities/InputParser.java
@@ -25,4 +25,48 @@ public class InputParser {
         return stringArray;
 
     }
+    
+    //WIP
+    public static Object parseParameter(String parameter) {
+
+	
+//	    	Cell currentCell = player.getCurrentCell(cellList);
+//	    	parameter = parameter.toLowerCase();
+//	
+//	    	ObjectComparator comp = new ObjectComparator();
+//	    	int notFound = -1;	    	
+//	    	
+//	    	int usableIndex = comp.isPresentUsable(currentCell.getItem(), usableList);
+//	    	int equipIndex = comp.isPresentEquipable(currentCell.getItem(), equipList);
+//	    	int keyIndex = comp.isPresentKeyItem(currentCell.getKeyItem(), keyList);
+//	
+//	    	if(usableIndex == notFound && equipIndex == notFound && keyIndex == notFound ) {
+//	    		// parameter not found
+//	    	}
+//	    	else {
+//	    		if (usableIndex > notFound) {
+//	
+//	    			Usable item = usableList.get(usableIndex);
+//	    			if (item.getName().toLowerCase().equals(parameter)){
+//	    				return item;
+//	    			}
+//	    		}
+//	    		if (equipIndex > notFound) {
+//	
+//	    			Equipable item = equipList.get(equipIndex);
+//	    			if (item.getName().toLowerCase().equals(parameter)){
+//	    				return item;
+//	    			}
+//	    		}
+//	    		if (keyIndex > notFound) {
+//	
+//	    			KeyItems item = keyList.get(keyIndex);
+//	    			if (item.getName().toLowerCase().equals(parameter)){
+//	    				return item;
+//	    			}
+//	    		}
+//	    	}
+//	
+	    	return null;
+    }
 }
